Refactoring and allowing distribution of non Remote objects in favor of Distributable Objects.
Basically this is so that users of CloudLauncher don't have to depend on java.rmi.Remote and the irritating RemoteExceptions that that interface enforces. This should make it easier to export regular old interfaces via Could Launch.

Also added a Distributor class which eventually should server as the single point of export for CloudLaunch distributed objects.