Setting userClassLoader for plugins to a new UserFirstClassLoader which gives user classes precedence.
This is needed because the PluginClassLoader is a child first classloader. In cases where a plugin share the same class name
space with the user classloader, this isn't always good when the plugin is trying to load user classes. The UserFirstClassLoader
gives precedence to user classes while still including the PluginClassLoader which gives plugins the option of doing more
traditional classloading when working with user objects.